refactor: move controllers to their own feature
